Moved shared portionso f CertificateIntegrationTestFixture to common test project
